// src/app/api/models/healthcare-provider.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Detalles del Healthcare Provider a crear.
 */
export interface HealthcareProvider {

  /**
   * Status of the provider
   */
  active?: boolean;

  /**
   * Official Provider Code
   */
  code: string;

  /**
   * Official Contact Email
   */
  email?: string;

  /**
   * Identificador único
   */
  id?: number;

  /**
   * Provider Name
   */
  name: string;

  /**
   * Official Contact Phone
   */
  phone?: string;

  /**
   * Tax Identification Number (NIT/RUT/DNI)
   */
  taxId: string;
}